Add a high level IO class
